Tidy up CommandList class. Split commands in CommandList into RecommendationsCommandList and ReviewCommandList. Move determineEditCommand into ReviewCommandList.
